+++ tests/test_cli.py
@@ -98,24 +98,6 @@ def test_django_proj(tmpdir):  # noqa
     cli.run()
 
 
-def test_with_numpydoc(tmpdir):  # noqa
-    sys.argv = ["pyscaffold", "--with-numpydoc", "my_project"]
-    cli.run()
-    conffile = os.path.join(
-        os.path.abspath(os.path.curdir), "my_project", "docs", "conf.py")
-    conf = imp.load_source('conf', conffile)
-    assert sorted(conf.extensions) == sorted(['sphinx.ext.autodoc',
-                                              'sphinx.ext.intersphinx',
-                                              'sphinx.ext.todo',
-                                              'sphinx.ext.autosummary',
-                                              'sphinx.ext.viewcode',
-                                              'sphinx.ext.coverage',
-                                              'sphinx.ext.doctest',
-                                              'sphinx.ext.ifconfig',
-                                              'sphinx.ext.pngmath',
-                                              'numpydoc'])
-
-
 def test_with_travis(tmpdir):  # noqa
     sys.argv = ["pyscaffold", "--with-travis", "my_project"]
     cli.run()